+++ apps/web/src/hooks/useExport.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
-import { useUnifiedStore } from '@/stores/useUnifiedStore';
-import type { ExportFormat } from '@/stores/unifiedStore';
+import { useUnifiedStore } from '@/hooks/useUnifiedStore';
+import type { ExportFormat } from '@/stores';
 
 export interface UseExportReturn {
   // Export operations
