+++ apps/web/src/stores/index.ts
@@ -1,13 +1,7 @@
-// Unified store (new architecture)
-export { 
-  useUnifiedStore,
-  useNodeById,
-  useArrowById,
-  usePersonById,
-  useSelectedEntity,
-  useIsExecuting,
-  useNodeExecutionState
-} from './useUnifiedStore';
+// Store exports only - hooks are in @/hooks
+export { useUnifiedStore } from './unifiedStore';
+
+// Type exports
 export type { 
   UnifiedStore,
   NodeState,
@@ -17,4 +11,7 @@ export type {
   ExportedArrow,
   ExportedPerson,
   ExportedApiKey
-} from './unifiedStore';
+} from './unifiedStore.types';
+
+// Export the DiagramExporter class
+export { DiagramExporter } from './diagramExporter';
